requirements.txt
awscli==1.42.66
botocore==1.40.66
Cartopy==0.25.0
certifi==2025.10.5
cftime==1.6.5
colorama==0.4.6
contourpy
cycler==0.12.1
docutils==0.19
fonttools==4.60.1
jmespath==1.0.1
kiwisolver==1.4.9
matplotlib==3.10.7
netCDF4==1.7.3
numpy>=1.21.4
packaging==25.0
pandas>=2.3.0
pillow==12.0.0
pyasn1==0.6.1
pyparsing==3.2.5
pyproj
pyshp==3.0.2.post1
python-dateutil>=2.8.0
pytz>=2022.7
PyYAML==6.0.3
rsa==4.7.2
s3transfer==0.14.0
shapely==2.1.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
xarray
kafka-python
s3fs
h5netcdf
pyarrow>=5.0.0
